refactor: 계산기 이외의 파일 디렉토리 분리
