+++ src/test/kotlin/calculator/StringCalculatorConvertTest.kt
@@ -1,5 +1,6 @@
 package calculator
 
+import calculator.model.CalculatorErrorCode
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.data.forAll
